% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colkurtosis}
\alias{colkurtosis}
\title{Compute column-wise kurtosis}
\usage{
colkurtosis(mtx)
}
\arguments{
\item{mtx}{a nxp matrix}
}
\value{
a 1xp vector
}
\description{
Compute column-wise kurtosis
}
