% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colsd}
\alias{colsd}
\title{Compute column-wise sd}
\usage{
colsd(mtx)
}
\arguments{
\item{mtx}{a nxp matrix}
}
\value{
a 1xp vector
}
\description{
Compute column-wise sd
}
\seealso{
Other column-wise and utilities: 
\code{\link{colkurtosis}()},
\code{\link{colskewness}()},
\code{\link{norm_minkowski}()}
}
\concept{column-wise and utilities}
