% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colskewness}
\alias{colskewness}
\title{Compute column-wise skewness}
\usage{
colskewness(mtx)
}
\arguments{
\item{mtx}{a nxp matrix}
}
\value{
a 1xp vector
}
\description{
Compute column-wise skewness
}
\seealso{
Other column-wise and utilities: 
\code{\link{colkurtosis}()},
\code{\link{colsd}()},
\code{\link{norm_minkowski}()}
}
\concept{column-wise and utilities}
