% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_distances_gen_obs_cpp}
\alias{compute_distances_gen_obs_cpp}
\title{Compute distances between summary statistics.}
\usage{
compute_distances_gen_obs_cpp(mtx_gen, mtx_obs, p_norm = 2)
}
\arguments{
\item{mtx_gen}{the generated data matrix}

\item{mtx_obs}{the observed data matrix}

\item{p_norm}{the power of the Minkowski distance}
}
\value{
a length-2 vector of distances between summary statistics
}
\description{
Compute distances between summary statistics.
}
