% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_acceptable_data_cpp}
\alias{generate_acceptable_data_cpp}
\title{Generate data that is accepted by ABC.}
\usage{
generate_acceptable_data_cpp(
  n_sample,
  m_sample,
  alpha_0,
  beta_0,
  nu_0,
  mtx_obs,
  summarize_eps,
  n_gen,
  max_iter,
  p_norm,
  use_optimized_summary
)
}
\arguments{
\item{n_sample}{hyperparameters that are used to generate data}

\item{m_sample}{hyperparameters that are used to generate data}

\item{alpha_0}{hyperparameters that are used to generate data}

\item{beta_0}{hyperparameters that are used to generate data}

\item{nu_0}{hyperparameters that are used to generate data}

\item{mtx_obs}{observed data}

\item{summarize_eps}{ABC threshold(s)}

\item{n_gen}{how many datasets are generated}

\item{max_iter}{how many iterations are tried}

\item{use_optimized_summary}{if TRUE, return the optimized summary statistics, else standard (mean, sd)}
}
\value{
a (n x n_obs x p) cube of generated data
}
\description{
Generate data that is accepted by ABC.
}
