% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_number_summary_statistics}
\alias{get_number_summary_statistics}
\title{Get the number of multivariate summary statistics.}
\usage{
get_number_summary_statistics(use_optimized_summary)
}
\arguments{
\item{use_optimized_summary}{if TRUE, return the optimized summary statistics (mean, sd, kurtosis, skewness), else standard (mean, sd)}
}
\value{
an integer
}
\description{
Get the number of multivariate summary statistics.
}
\seealso{
Other ABC summary functions: 
\code{\link{get_optimized_summary_statistics_cpp}()},
\code{\link{get_standard_summary_statistics_cpp}()},
\code{\link{get_summary_statistics_cpp}()}
}
\concept{ABC summary functions}
