% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_minkowski}
\alias{norm_minkowski}
\title{Compute the Minkowski norm of a vector}
\usage{
norm_minkowski(v, p = 2)
}
\arguments{
\item{v}{a vector}

\item{p}{exponent of the Minkowski norm (from 1 to Inf)}
}
\value{
a double
}
\description{
Compute the Minkowski norm of a vector.
$p$ can range from 1 to infinity.
}
