% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdirdirgamma_beta_cpp}
\alias{rdirdirgamma_beta_cpp}
\title{Generate a Dirichlet-Dirichlet-Gamma population (safer).}
\usage{
rdirdirgamma_beta_cpp(n, m, alpha_0, beta_0, nu_0)
}
\arguments{
\item{n}{number of samples per source}

\item{m}{number of sources}

\item{alpha_0}{between-source Gamma hyperparameter, a scalar}

\item{beta_0}{between-source Gamma hyperparameter, a scalar}

\item{nu_0}{between-source Dirichlet hyperparameter, a numeric vector}
}
\value{
a matrix with n*m rows
}
\description{
Generate samples from m sources and p parameters, n sample per source.
The between-source alpha hyperparameter used to generate the source parameters is mandatory.
}
\details{
\subsection{RNG}{

this function uses R's RNG seed.
}
}
\seealso{
Other RNG functions: 
\code{\link{rdirdirgamma_cpp}()},
\code{\link{rdirichlet_beta_cpp}()},
\code{\link{rdirichlet_cpp}()}
}
\concept{RNG functions}
