% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdirdirgamma_cpp}
\alias{rdirdirgamma_cpp}
\title{Generate a Dirichlet-Dirichlet-Gamma population (unsafe).}
\usage{
rdirdirgamma_cpp(n, m, alpha_0, beta_0, nu_0, seed = 0L)
}
\arguments{
\item{n}{number of samples per source}

\item{m}{number of sources}

\item{alpha_0}{between-source Gamma hyperparameter, a scalar}

\item{beta_0}{between-source Gamma hyperparameter, a scalar}

\item{nu_0}{between-source alpha hyperparameter, a numeric vector.}
}
\description{
Generate samples from m sources and p parameters, n sample per source.
The between-source alpha hyperparameter used to generate the source parameters is mandatory.
}
