% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdirichlet_beta_cpp}
\alias{rdirichlet_beta_cpp}
\title{Generate from a Dirichlet distribution using the stick breaking definition (safer).}
\usage{
rdirichlet_beta_cpp(n, alpha)
}
\arguments{
\item{n}{how many samples to generate}

\item{alpha}{the Dirichlet hyperparameter, with p entries}
}
\value{
a numeric matrix, n*p
}
\description{
Generate from a Dirichlet distribution using the stick breaking definition (safer).
}
