% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdirichlet_cpp}
\alias{rdirichlet_cpp}
\title{Generate one sample from a Dirichlet distribution.}
\usage{
rdirichlet_cpp(alpha, seed = 0L)
}
\arguments{
\item{alpha}{the Dirichlet hyperparameter}

\item{seed}{the RNG seed: if 0 (default), generate a time-based seed}
}
\value{
a numeric vector
}
\description{
\subsection{RNG}{
}
}
\details{
this function uses GSL's RNG seed, unaffected by R's RNG.
}
