% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_ABC_cpp}
\alias{compute_ABC_cpp}
\title{Perform ABC sampling using the stick breaking procedure, returning the acceptance ratio.}
\usage{
compute_ABC_cpp(
  n_sample,
  m_sample,
  alpha_0,
  beta_0,
  nu_0,
  mtx_obs,
  summarize_eps,
  max_iter,
  p_norm,
  use_optimized_summary
)
}
\arguments{
\item{n_sample}{hyperparameters that are used to generate data}

\item{m_sample}{hyperparameters that are used to generate data}

\item{alpha_0}{hyperparameters that are used to generate data}

\item{beta_0}{hyperparameters that are used to generate data}

\item{nu_0}{hyperparameters that are used to generate data}

\item{mtx_obs}{observed data}

\item{summarize_eps}{ABC thresholds: as many as summary statistics}

\item{max_iter}{how many iterations are tried}

\item{p_norm}{exponent of the L^p norm (can be `Inf`)}

\item{use_optimized_summary}{if TRUE, return the optimized summary statistics, else standard (mean, sd)}
}
\value{
the acceptance ratio, where 1 means that all max_iter samples were accepted.
}
\description{
Perform ABC sampling using the stick breaking procedure, returning the acceptance ratio.
}
