% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_ABC_cpp}
\alias{compute_ABC_cpp}
\title{Perform ABC sampling using the stick breaking procedure, returning the acceptance ratio.}
\usage{
compute_ABC_cpp(
  n_sample,
  m_sample,
  alpha_0,
  beta_0,
  nu_0,
  mtx_obs,
  summarize_eps,
  reps,
  p_norm,
  use_optimized_summary,
  return_distances = FALSE
)
}
\arguments{
\item{n_sample}{hyperparameters that are used to generate data: number of samples per source}

\item{m_sample}{hyperparameters that are used to generate data: number of sources}

\item{alpha_0}{hyperparameters that are used to generate data}

\item{beta_0}{hyperparameters that are used to generate data}

\item{nu_0}{hyperparameters that are used to generate data}

\item{mtx_obs}{the observed data matrix}

\item{summarize_eps}{ABC thresholds: as many as summary statistics}

\item{reps}{number of ABC samples to generate}

\item{p_norm}{exponent of the L^p norm (can be \code{Inf}) (default: 2)}

\item{use_optimized_summary}{if TRUE, return the optimized summary statistics (mean, sd, skewness, kurtosis), else standard (mean, sd)}

\item{return_distances}{if TRUE, also return distances for all samples}
}
\value{
if \code{return_distances} is FALSE, a list with the acceptance ratio, where 1 means that all max_iter samples were accepted.
if \code{return_distances} is TRUE, a list, with components \code{d_ABC} (a max_iter x n_summary matrix) and the acceptance ratio
}
\description{
Perform ABC sampling using the stick breaking procedure, returning the acceptance ratio.
Similar to \code{\link[=sample_ABC_rdirdirgamma_beta_cpp]{sample_ABC_rdirdirgamma_beta_cpp()}} but also performs the acceptance step.
}
