% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_distances_gen_obs_cpp}
\alias{compute_distances_gen_obs_cpp}
\title{Compute distances between summary statistics.}
\usage{
compute_distances_gen_obs_cpp(
  mtx_gen,
  mtx_obs,
  p_norm = 2,
  use_optimized_summary = FALSE
)
}
\arguments{
\item{mtx_gen}{the generated data matrix; number of rows is free, it must have the same number of columns as \code{mtx_obs}}

\item{mtx_obs}{the observed data matrix; number of rows is free, it must have the same number of columns as \code{mtx_gen}}

\item{p_norm}{exponent of the L^p norm (can be \code{Inf}) (default: 2)}

\item{use_optimized_summary}{if TRUE, use quantile matrix, else compute mean and sd vectors}
}
\value{
a vector of distances between summary statistics: as many entries as summary statistics
}
\description{
Compute distances between summary statistics.
}
\seealso{
Other ABC functions: 
\code{\link{compute_ABC_cpp}()},
\code{\link{generate_acceptable_data_cpp}()},
\code{\link{sample_ABC_rdirdirgamma_beta_cpp}()}
}
\concept{ABC functions}
