% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_acceptable_data_cpp}
\alias{generate_acceptable_data_cpp}
\title{Generate data that is accepted by ABC.}
\usage{
generate_acceptable_data_cpp(
  n_sample,
  m_sample,
  alpha_0,
  beta_0,
  nu_0,
  mtx_obs,
  summarize_eps,
  reps,
  max_iter,
  p_norm,
  use_optimized_summary
)
}
\arguments{
\item{n_sample}{hyperparameters that are used to generate data: number of samples per source}

\item{m_sample}{hyperparameters that are used to generate data: number of sources}

\item{alpha_0}{hyperparameters that are used to generate data}

\item{beta_0}{hyperparameters that are used to generate data}

\item{nu_0}{hyperparameters that are used to generate data}

\item{mtx_obs}{the observed data matrix}

\item{summarize_eps}{ABC thresholds: as many as summary statistics}

\item{reps}{how many datasets are returned}

\item{max_iter}{how many iterations are tried}

\item{p_norm}{exponent of the L^p norm (can be \code{Inf}) (default: 2)}

\item{use_optimized_summary}{if TRUE, return the optimized summary statistics (mean, sd, skewness, kurtosis), else standard (mean, sd)}
}
\value{
a (n x n_obs x p) array of generated data
}
\description{
Generate data that is accepted by ABC.
}
\seealso{
Other ABC functions: 
\code{\link{compute_ABC_cpp}()},
\code{\link{compute_distances_gen_obs_cpp}()},
\code{\link{sample_ABC_rdirdirgamma_beta_cpp}()}
}
\concept{ABC functions}
