% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_number_summary_statistics}
\alias{get_number_summary_statistics}
\title{Get the number of multivariate summary statistics.}
\usage{
get_number_summary_statistics(use_optimized_summary)
}
\value{
an integer
}
\description{
Get the number of multivariate summary statistics.
}
