% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_optimized_summary_statistics_cpp}
\alias{get_optimized_summary_statistics_cpp}
\title{Compute optimized summary statistics.}
\usage{
get_optimized_summary_statistics_cpp(mtx)
}
\arguments{
\item{mtx}{a data matrix (nxp)}
}
\value{
a kxp matrix of summary statistics
}
\description{
Compute optimized summary statistics.
}
