% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_standard_summary_statistics_cpp}
\alias{get_standard_summary_statistics_cpp}
\title{Compute standard summary statistics.}
\usage{
get_standard_summary_statistics_cpp(mtx)
}
\arguments{
\item{mtx}{a data matrix (nxp)}
}
\value{
a kxp matrix of summary statistics
}
\description{
Compute standard summary statistics:
}
\details{
\itemize{
\item column-wise mean
\item column-wise standard deviation
}
}
\seealso{
Other ABC summary functions: 
\code{\link{get_number_summary_statistics}()},
\code{\link{get_optimized_summary_statistics_cpp}()},
\code{\link{get_summary_statistics_cpp}()}
}
\concept{ABC summary functions}
