% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdirdirgamma_cpp}
\alias{rdirdirgamma_cpp}
\title{Generate a Dirichlet-Dirichlet-Gamma population (unsafe).}
\usage{
rdirdirgamma_cpp(n, m, alpha_0, beta_0, nu_0, seed = 0L)
}
\arguments{
\item{n}{number of samples per source}

\item{m}{number of sources}

\item{alpha_0}{between-source Gamma hyperparameter, a scalar}

\item{beta_0}{between-source Gamma hyperparameter, a scalar}

\item{nu_0}{between-source Dirichlet hyperparameter, a numeric vector}

\item{seed}{the RNG seed: if 0 (default), generate a time-based seed}
}
\value{
a matrix with n*m rows
}
\description{
Generate samples from m sources and p parameters, n sample per source.
The between-source alpha hyperparameter used to generate the source parameters is mandatory.
}
\details{
## RNG

this function uses GSL's RNG seed, unaffected by R's RNG.
}
