% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_ABC_rdirdirgamma_beta_cpp}
\alias{sample_ABC_rdirdirgamma_beta_cpp}
\title{Perform ABC sampling and distance calculation using the stick breaking procedure.}
\usage{
sample_ABC_rdirdirgamma_beta_cpp(
  n,
  m,
  alpha_0,
  beta_0,
  nu_0,
  mtx_obs,
  reps,
  p_norm = 2L
)
}
\arguments{
\item{n}{passed to [rdirdirgamma_cpp()]}

\item{m}{passed to [rdirdirgamma_cpp()]}

\item{alpha_0}{hyperparameter}

\item{beta_0}{hyperparameter}

\item{nu_0}{hyperparameter}

\item{mtx_obs}{the observed data matrix}

\item{reps}{repetitions to average distances (default: 1)}

\item{method}{passed to [stats::dist()]}
}
\description{
Perform ABC sampling and distance calculation using the stick breaking procedure.
}
