% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_ABC_rdirdirgamma_beta_cpp}
\alias{sample_ABC_rdirdirgamma_beta_cpp}
\title{Perform ABC sampling and distance calculation using the stick breaking procedure.}
\usage{
sample_ABC_rdirdirgamma_beta_cpp(
  n_sample,
  m_sample,
  alpha_0,
  beta_0,
  nu_0,
  mtx_obs,
  reps,
  p_norm = 2
)
}
\arguments{
\item{n_sample}{number of samples per source}

\item{m_sample}{number of sources}

\item{mtx_obs}{the observed data matrix}

\item{reps}{number of ABC samples (default: 1)}

\item{p_norm}{exponent of the L^p norm (can be `Inf`) (default: `2`)}
}
\value{
a reps*2 matrix of distances between summary statistics
}
\description{
Samples from Dirichlet using [rdirdirgamma_beta_cpp()].
}
\details{
Summary statistics between datasets:

- mean
- standard deviation
}
