% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_ABC_rdirdirgamma_beta_cpp}
\alias{sample_ABC_rdirdirgamma_beta_cpp}
\title{Perform ABC sampling and distance calculation using the stick breaking procedure.}
\usage{
sample_ABC_rdirdirgamma_beta_cpp(
  n_sample,
  m_sample,
  alpha_0,
  beta_0,
  nu_0,
  mtx_obs,
  reps,
  p_norm = 2,
  use_optimized_summary = FALSE
)
}
\arguments{
\item{n_sample}{hyperparameters that are used to generate data: number of samples per source}

\item{m_sample}{hyperparameters that are used to generate data: number of sources}

\item{alpha_0}{hyperparameters that are used to generate data}

\item{beta_0}{hyperparameters that are used to generate data}

\item{nu_0}{hyperparameters that are used to generate data}

\item{mtx_obs}{the observed data matrix}

\item{reps}{number of ABC samples to generate}

\item{p_norm}{exponent of the L^p norm (can be \code{Inf}) (default: 2)}

\item{use_optimized_summary}{if TRUE, return the optimized summary statistics (mean, sd, skewness, kurtosis), else standard (mean, sd)}
}
\value{
a reps*2 matrix of distances between summary statistics
}
\description{
Perform ABC sampling and distance calculation using the stick breaking procedure.
}
\details{
\subsection{Procedure}{
\enumerate{
\item samples from Dirichlet using \code{\link[=rdirdirgamma_beta_cpp]{rdirdirgamma_beta_cpp()}}.
\item computes summary statistics on datasets:
}
\itemize{
\item column-wise mean
\item column-wise standard deviation
}

or a set of column-wise moments (mean, sd, skewness, kurtosis).
\enumerate{
\item the generated dataset is invisibly is truncated to the same amount of rows as the observed dataset.
\item compute the Minkowski norms of the differences between summary statistics.
\item repeat \code{reps} times.
}
}

\subsection{RNG}{

this function uses R's RNG seed.
}
}
\seealso{
Other ABC functions: 
\code{\link{compute_ABC_cpp}()},
\code{\link{compute_distances_gen_obs_cpp}()},
\code{\link{generate_acceptable_data_cpp}()}
}
\concept{ABC functions}
